Use AssemlyInfo version to set window title
Going forward, bumping up version will require a change to the
`AssemblyInfo.cs` file, but it won't require changing the form/window
title, as that is derived from the version there. For people trying to
check the code, it will be more obvious where to look for the current
version, as previously this was only exposed as the hardcoded form
title.