Move event event handlers setup to designer
After moving the handles to the designer / `InitializeComponent`, we
need to ensure, that anything needed by the setting changed handlers is
initialized too before loading and setting the checked states.

For now this only applies to AutoBorderless - after setting Checked to
true, it tries to start the process monitor, so `ProcessMonitor` must be
initialized before that.